# DF/src/erd/vgph_system_setup/vgph_system_setup.controller.ts
import { Controller, Get, Post, Body, Patch, Param, Delete,UseGuards,Query,Req,NotFoundException,Headers} from '@nestjs/common';
import { vgph_system_setupService } from './vgph_system_setup.service';
import { Prisma } from '@prisma/client';
import { ApiOkResponse, ApiTags,ApiOperation,ApiBody,
  ApiQuery,ApiParam,
  ApiBadRequestResponse,ApiUnauthorizedResponse,
  ApiForbiddenResponse,ApiNotAcceptableResponse,
  ApiConflictResponse,ApiNotFoundResponse,
  ApiMethodNotAllowedResponse,
  ApiRequestTimeoutResponse,
  ApiGoneResponse,
  ApiUnsupportedMediaTypeResponse,
  ApiUnprocessableEntityResponse,
  ApiInternalServerErrorResponse,
  ApiNotImplementedResponse,
  ApiBadGatewayResponse,
  ApiServiceUnavailableResponse,
  ApiGatewayTimeoutResponse,
  ApiBearerAuth 
} from '@nestjs/swagger';
import { vgph_system_setupEntity } from './entity/vgph_system_setup.entity';
//import { CreateVgphSystemSetupDto } from '../prisma/dto/create-vgphSystemSetup.dto';
//import { UpdateVgphSystemSetupDto } from '../prisma/dto/update-vgphSystemSetup.dto';
import { Createvgph_system_setupDto } from './dto/Createvgph_system_setup.dto';
import { Updatevgph_system_setupDto } from './dto/Updatevgph_system_setup.dto';
import { plainToInstance } from 'class-transformer';
import { UfService } from 'src/Torus/v4/uf/uf.service';

 
@Controller('vgph_system_setup')
@ApiTags('ERD API')
export class vgph_system_setupController {
  constructor(
    private readonly vgph_system_setupService: vgph_system_setupService,
    private readonly ufservice: UfService
  ) {}

  @Get("/schema")
  @ApiBearerAuth('JWT-auth')
  @ApiOkResponse({ type: vgph_system_setupEntity })
  @ApiOperation({
    summary: 'schema validation',
    description: 'Retrive the datatype of the vgph_system_setup table',
  })
  async findSchema(@Headers() authHeader: string,@Req() req: any) {
    const token = req.headers?.authorization?.split(' ')[1];
    //await this.ufservice.introspectToken(authHeader,"",token);
    return this.vgph_system_setupService.findSchema(token);
  }

  @Get('/get')
  @ApiBearerAuth('JWT-auth')
  @ApiOkResponse({ type: vgph_system_setupEntity, isArray: true })
  @ApiOperation({
    summary: 'Filter the records',
    description: 'Filter all the records from the vgph_system_setup table',
  })
  @ApiQuery({ name: 'limit', required: false, type: Number, description: 'Number of records to fetch' })
  async findAllmethod(@Headers() authHeader: string,@Query() query: any,@Body() body: any,@Req() req: any) {
    const token = req.headers?.authorization?.split(' ')[1];
    //await this.ufservice.introspectToken(authHeader,"",token);
    const { limit }:{ limit:number } = query;
    const { selectColumns}:{ selectColumns:any } = body;
    return this.vgph_system_setupService.findAllmethod(query, +limit,selectColumns,token);
  }

  @Get(':vgph_id')
  @ApiBearerAuth('JWT-auth')
  @ApiParam({name: 'vgph_id',type:Number})
  @ApiOkResponse({ type: vgph_system_setupEntity })
  @ApiOperation({
    summary: 'Fetch the only one record',
    description: 'Read only one records from the vgph_system_setup table',
  })
  
  async findOne(@Headers() authHeader: string,@Param('vgph_id') vgph_id:number,@Req() req: any) {
    const token = req.headers?.authorization?.split(' ')[1];
    //await this.ufservice.introspectToken(authHeader,"",token);
    const result = this.vgph_system_setupService.findOne(+vgph_id,token);
    return plainToInstance(vgph_system_setupEntity, result);
  }
 
  @Get()
  @ApiBearerAuth('JWT-auth')
  @ApiQuery({ name: 'vgph_id', required: false, type: Number})
  @ApiOkResponse({ type: vgph_system_setupEntity, isArray: true })
  @ApiOperation({
    summary: 'Read all the records',
    description: 'Read all the records from the vgph_system_setup table',
  })
  
  async findAll(@Headers() authHeader: string,@Req() req: any,@Query("trs_creator_email") trs_creator_email?: string,@Query("trs_created_date") trs_created_date?: Date,@Query("trs_created_by") trs_created_by?: string,@Query("trs_modified_date") trs_modified_date?: Date,@Query("trs_modified_by") trs_modified_by?: string,@Query("trs_next_status") trs_next_status?: string,@Query("trs_status") trs_status?: string,@Query("trs_process_id") trs_process_id?: string,@Query("trs_access_profile") trs_access_profile?: string,@Query("trs_org_grp_code") trs_org_grp_code?: string,@Query("trs_org_code") trs_org_code?: string,@Query("trs_role_grp_code") trs_role_grp_code?: string,@Query("trs_role_code") trs_role_code?: string,@Query("trs_ps_grp_code") trs_ps_grp_code?: string,@Query("trs_ps_code") trs_ps_code?: string,@Query("trs_sub_org_grp_code") trs_sub_org_grp_code?: string,@Query("trs_sub_org_code") trs_sub_org_code?: string,@Query('vgph_id') vgph_id?:string,@Query() query?: Record<string, any>) {
    const token = req.headers?.authorization?.split(' ')[1];
    //await this.ufservice.introspectToken(authHeader,"",token);
    let presentQueryKeys:any=[
      'vgph_id',
      "trs_creator_email",
      "trs_created_date",
      "trs_created_by",
      "trs_modified_date",
      "trs_modified_by",
      "trs_next_status",
      "trs_status",
      "trs_process_id",
      "trs_access_profile",
      "trs_org_grp_code",
      "trs_org_code",
      "trs_role_grp_code",
      "trs_role_code",
      "trs_ps_grp_code",
      "trs_ps_code",
      "trs_sub_org_grp_code",
      "trs_sub_org_code"
    ]
    let comingQueryKeys:any=Object.keys(query)||[]
    let isComingQuerysAreValid=true;
    if(comingQueryKeys.length==0)
      {
        isComingQuerysAreValid = true;
      }
  
      // If arrays have different lengths, they cannot be equal
      if (comingQueryKeys.length > presentQueryKeys.length) {
        isComingQuerysAreValid= false;
      }
      // Compare each element after sorting
      for (let i = 0; i < comingQueryKeys.length; i++) {
        if (!presentQueryKeys.includes(comingQueryKeys[i])) {
          isComingQuerysAreValid=false;
        }
      }
    if (req.originalUrl.includes('?') && req.originalUrl.split('?')[1].includes('/') || isComingQuerysAreValid==false) {
      throw new NotFoundException('Invalid query parameter structure.');
    }
    const result = this.vgph_system_setupService.findAll(token,trs_creator_email,trs_created_date,trs_created_by,trs_modified_date,trs_modified_by,trs_next_status,trs_status,trs_process_id,trs_access_profile,trs_org_grp_code,trs_org_code,trs_role_grp_code,trs_role_code,trs_ps_grp_code,trs_ps_code,trs_sub_org_grp_code,trs_sub_org_code,+vgph_id);
    return plainToInstance(vgph_system_setupEntity, result);
  } 

  @Post()
  @ApiBearerAuth('JWT-auth')
  @ApiBody({ type: Createvgph_system_setupDto })
  @ApiOkResponse({ type: vgph_system_setupEntity })
  @ApiOperation({
    summary: 'Create the record',
    description: 'Create the record for the vgph_system_setup table',
  })
  
  async create(@Headers() authHeader: string,@Body() createvgph_system_setupDto: Prisma.vgph_system_setupCreateInput,@Req() req: any) {
    const token = req.headers?.authorization?.split(' ')[1];
    //await this.ufservice.introspectToken(authHeader,"",token);
    const result = this.vgph_system_setupService.create(createvgph_system_setupDto,token);
    return plainToInstance(vgph_system_setupEntity, result);
  }
 
  @Patch(':vgph_id')
  @ApiBearerAuth('JWT-auth')
  @ApiParam({name: 'vgph_id',type:Number})
  @ApiBody({ type: Updatevgph_system_setupDto })
  @ApiOkResponse({ type: vgph_system_setupEntity })
  @ApiOperation({
    summary: 'Update the record',
    description: 'Update the record for the vgph_system_setup table',
  })
    
  async update(@Headers() authHeader: string,@Param('vgph_id') vgph_id:number,
    @Body() updatevgph_system_setupDto: Prisma.vgph_system_setupUpdateInput,
    @Req() req: any) {
    const token = req.headers?.authorization?.split(' ')[1];
    //await this.ufservice.introspectToken(authHeader,"",token);
    const result = this.vgph_system_setupService.update(+vgph_id,updatevgph_system_setupDto,token);
    return plainToInstance(vgph_system_setupEntity, result);
  }
 
  @Delete(':vgph_id')
  @ApiBearerAuth('JWT-auth')
  @ApiParam({name: 'vgph_id',type:Number})
  @ApiOkResponse({ type: vgph_system_setupEntity })
  @ApiOperation({
    summary: 'Delete the record',
    description: 'Delete the record for the vgph_system_setup table',
  })
  
  async remove(@Headers() authHeader: string,@Param('vgph_id') vgph_id:number,@Req() req: any) {
    const token = req.headers?.authorization?.split(' ')[1];
    //await this.ufservice.introspectToken(authHeader,"",token);
    const result =  this.vgph_system_setupService.remove(+vgph_id,token);
    return plainToInstance(vgph_system_setupEntity, result);
  }  
 
  @Get('/find/first')
  @ApiBearerAuth('JWT-auth')
  //@ApiParam({name: ''})
  @ApiOkResponse({ type: vgph_system_setupEntity })
  @ApiOperation({
    summary: 'Fetch the first record',
    description: 'Read first record from the vgph_system_setup table',
  })
  
  async findFirst(@Headers() authHeader: string,@Param() params: any, @Req() req: any) {
    const token = req.headers?.authorization?.split(' ')[1];
    //await this.ufservice.introspectToken(authHeader,"",token);
    const result = this.vgph_system_setupService.findFirst(token);
    return plainToInstance(vgph_system_setupEntity, result);
  }

  @Get('/find/last')
  @ApiBearerAuth('JWT-auth')
  //@ApiParam({name: ''})
  @ApiOkResponse({ type: vgph_system_setupEntity })
  @ApiOperation({
    summary: 'Fetch the last record',
    description: 'Read last record from the vgph_system_setup table',
  })
  
  async findLast(@Headers() authHeader: string,@Param() params: any, @Req() req: any) {
    const token = req.headers?.authorization?.split(' ')[1];
    //await this.ufservice.introspectToken(authHeader,"",token);
    const result = this.vgph_system_setupService.findLast(token);
    return plainToInstance(vgph_system_setupEntity, result);
  }
}