# DF/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/vgph_system_setup/schema": {
      "get": {
        "description": "Retrive the datatype of the vgph_system_setup table",
        "operationId": "vgph_system_setupController_findSchema",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/vgph_system_setupEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "schema validation",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/vgph_system_setup/get": {
      "get": {
        "description": "Filter all the records from the vgph_system_setup table",
        "operationId": "vgph_system_setupController_findAllmethod",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to fetch",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/vgph_system_setupEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Filter the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/vgph_system_setup/{vgph_id}": {
      "get": {
        "description": "Read only one records from the vgph_system_setup table",
        "operationId": "vgph_system_setupController_findOne",
        "parameters": [
          {
            "name": "vgph_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/vgph_system_setupEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the only one record",
        "tags": [
          "ERD API"
        ]
      },
      "patch": {
        "description": "Update the record for the vgph_system_setup table",
        "operationId": "vgph_system_setupController_update",
        "parameters": [
          {
            "name": "vgph_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Updatevgph_system_setupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/vgph_system_setupEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update the record",
        "tags": [
          "ERD API"
        ]
      },
      "delete": {
        "description": "Delete the record for the vgph_system_setup table",
        "operationId": "vgph_system_setupController_remove",
        "parameters": [
          {
            "name": "vgph_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/vgph_system_setupEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/vgph_system_setup": {
      "get": {
        "description": "Read all the records from the vgph_system_setup table",
        "operationId": "vgph_system_setupController_findAll",
        "parameters": [
          {
            "name": "trs_creator_email",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_created_date",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "trs_created_by",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_modified_date",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "trs_modified_by",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_next_status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_process_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_access_profile",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_org_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_org_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_role_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_role_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_ps_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_ps_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_sub_org_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_sub_org_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vgph_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/vgph_system_setupEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "post": {
        "description": "Create the record for the vgph_system_setup table",
        "operationId": "vgph_system_setupController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Createvgph_system_setupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/vgph_system_setupEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/vgph_system_setup/find/first": {
      "get": {
        "description": "Read first record from the vgph_system_setup table",
        "operationId": "vgph_system_setupController_findFirst",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/vgph_system_setupEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the first record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/vgph_system_setup/find/last": {
      "get": {
        "description": "Read last record from the vgph_system_setup table",
        "operationId": "vgph_system_setupController_findLast",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/vgph_system_setupEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the last record",
        "tags": [
          "ERD API"
        ]
      }
    }
  },
  "info": {
    "title": "Nest API",
    "description": "The Nestjs API documentation",
    "version": "0.4",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://192.168.2.57:3010",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "vgph_system_setupEntity": {
        "type": "object",
        "properties": {
          "setup_code": {
            "type": "string",
            "example": "string"
          },
          "product_code": {
            "type": "string",
            "example": "string"
          },
          "interface_product": {
            "type": "string",
            "example": "string"
          },
          "category": {
            "type": "string",
            "example": "string"
          },
          "sub_category": {
            "type": "string",
            "example": "string"
          },
          "purpose": {
            "type": "string",
            "example": "string"
          },
          "setup_value": {
            "type": "object",
            "example": "any"
          },
          "vgph_id": {
            "type": "number",
            "example": "number"
          },
          "trs_creator_email": {
            "type": "string",
            "example": "string"
          },
          "trs_created_date": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_created_by": {
            "type": "string",
            "example": "string"
          },
          "trs_modified_date": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_modified_by": {
            "type": "string",
            "example": "string"
          },
          "trs_status": {
            "type": "string",
            "example": "string"
          },
          "trs_next_status": {
            "type": "string",
            "example": "string"
          },
          "trs_process_id": {
            "type": "string",
            "example": "string"
          },
          "trs_access_profile": {
            "type": "string",
            "example": "string"
          },
          "trs_org_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_org_code": {
            "type": "string",
            "example": "string"
          },
          "trs_role_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_role_code": {
            "type": "string",
            "example": "string"
          },
          "trs_ps_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_ps_code": {
            "type": "string",
            "example": "string"
          },
          "trs_sub_org_code": {
            "type": "string",
            "example": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "setup_code",
          "product_code",
          "interface_product",
          "category",
          "sub_category",
          "purpose",
          "setup_value",
          "vgph_id",
          "trs_creator_email",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code",
          "trs_sub_org_code",
          "trs_sub_org_grp_code"
        ]
      },
      "Createvgph_system_setupDto": {
        "type": "object",
        "properties": {
          "setup_code": {
            "type": "string"
          },
          "product_code": {
            "type": "string"
          },
          "interface_product": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "sub_category": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "setup_value": {
            "type": "object"
          },
          "trs_creator_email": {
            "type": "string"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string"
          },
          "trs_sub_org_code": {
            "type": "string"
          }
        },
        "required": [
          "setup_code",
          "product_code",
          "interface_product",
          "category",
          "sub_category",
          "purpose",
          "setup_value",
          "trs_creator_email",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code",
          "trs_sub_org_grp_code",
          "trs_sub_org_code"
        ]
      },
      "Updatevgph_system_setupDto": {
        "type": "object",
        "properties": {
          "setup_code": {
            "type": "string"
          },
          "product_code": {
            "type": "string"
          },
          "interface_product": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "sub_category": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "setup_value": {
            "type": "object"
          },
          "trs_creator_email": {
            "type": "string",
            "example": "string"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          },
          "trs_sub_org_code": {
            "type": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string"
          }
        },
        "required": [
          "setup_code",
          "product_code",
          "interface_product",
          "category",
          "sub_category",
          "purpose",
          "setup_value",
          "trs_creator_email",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code",
          "trs_sub_org_code",
          "trs_sub_org_grp_code"
        ]
      }
    }
  }
}